refactor: rename AgentCallbacks to AgentProgressCallbacks and move RichCallbacks to top-level as PrintAgentProgressCallbacks
