+++ src/coding_assistant/agents/callbacks.py
@@ -6,17 +6,8 @@ from abc import ABC, abstractmethod
 from pprint import pformat
 from typing import Any
 
-from rich import print
-from rich.console import Group
-from rich.json import JSON
-from rich.markdown import Markdown
-from rich.padding import Padding
-from rich.panel import Panel
-from rich.pretty import Pretty
-from rich.text import Text
 
-
-class AgentCallbacks(ABC):
+class AgentProgressCallbacks(ABC):
     """Abstract interface for agent callbacks."""
 
     @abstractmethod
@@ -60,7 +51,7 @@ class AgentCallbacks(ABC):
         pass
 
 
-class NullCallbacks(AgentCallbacks):
+class NullCallbacks(AgentProgressCallbacks):
     """Null object implementation that does nothing."""
 
     def on_agent_start(self, agent_name: str, model: str, is_resuming: bool = False):
@@ -86,95 +77,3 @@ class NullCallbacks(AgentCallbacks):
 
     def on_chunks_end(self):
         pass
-
-
-class RichCallbacks(AgentCallbacks):
-    def __init__(self, print_chunks: bool = True, print_reasoning: bool = True):
-        self._print_chunks = print_chunks
-        self._print_reasoning = print_reasoning
-
-    def on_agent_start(self, agent_name: str, model: str, is_resuming: bool = False):
-        status = "resuming" if is_resuming else "starting"
-        print(
-            Panel(
-                "",
-                title=f"Agent {agent_name} ({model}) {status}",
-                border_style="red",
-            ),
-        )
-
-    def on_agent_end(self, agent_name: str, result: str, summary: str):
-        quoted_result = textwrap.indent(result, "> ", lambda _: True)
-        quoted_summary = textwrap.indent(summary, "> ", lambda _: True)
-        print(
-            Panel(
-                Markdown(f"Result\n\n{quoted_result}\n\nSummary\n\n{quoted_summary}"),
-                title=f"Agent {agent_name} result",
-                border_style="red",
-            ),
-        )
-
-    def on_user_message(self, agent_name: str, content: str):
-        print(
-            Panel(
-                Markdown(content),
-                title=f"Agent {agent_name} user",
-                border_style="blue",
-            ),
-        )
-
-    def on_assistant_message(self, agent_name: str, content: str):
-        print(
-            Panel(
-                Markdown(content),
-                title=f"Agent {agent_name} assistant",
-                border_style="green",
-            ),
-        )
-
-    def on_assistant_reasoning(self, agent_name: str, content: str):
-        if self._print_reasoning:
-            print(
-                Panel(
-                    Markdown(content),
-                    title=f"Agent {agent_name} reasoning",
-                    border_style="cyan",
-                ),
-            )
-
-    def _try_parse_json(self, content: str):
-        try:
-            return json.loads(content)
-        except json.JSONDecodeError:
-            return None
-
-    def _format_tool_result(self, result: str):
-        if data := self._try_parse_json(result):
-            return Pretty(data, expand_all=True, indent_size=2)
-        else:
-            return Markdown(f"```\n{result}\n```")
-
-    def on_tool_message(self, agent_name: str, tool_name: str, arguments: dict | None, result: str):
-        parts: list[Any] = [Markdown(f"Name: `{tool_name}`")]
-
-        if arguments is not None:
-            parts.append(Padding(Pretty(arguments, expand_all=True, indent_size=2), (1, 0, 0, 0)))
-
-        parts.append(Padding(self._format_tool_result(result), (1, 0, 0, 0)))
-
-        render_group = Group(*parts)
-        print(
-            Panel(
-                render_group,
-                title=f"Agent {agent_name} tool call",
-                border_style="yellow",
-            ),
-        )
-
-    def on_chunk(self, chunk: str):
-        if self._print_chunks:
-            print(chunk, end="", flush=True)
-
-    def on_chunks_end(self):
-        if self._print_chunks:
-            print()
