+++ src/coding_assistant/agents/types.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from dataclasses import dataclass, field
 from typing import Awaitable, Protocol
 
-from coding_assistant.agents.callbacks import AgentCallbacks
+from coding_assistant.agents.callbacks import AgentProgressCallbacks
 from coding_assistant.agents.parameters import Parameter
 from coding_assistant.llm.model import Completion
 
@@ -86,5 +86,5 @@ class Completer(Protocol):
         *,
         model: str,
         tools: list,
-        callbacks: AgentCallbacks,
+    callbacks: AgentProgressCallbacks,
     ) -> Awaitable[Completion]: ...
